its coming to the end
We're leaving together
But still it's farewell
And maybe we'll come back
To earth, who can tell?
I guess there is no one to blame
We're leaving ground (leaving ground)
Will things ever be the same again?
[Hook]
It's the final countdown
The final countdown
[Verse 2]
We're heading for Venus (Venus)
And still we stand tall
Cause maybe they've seen us
And welcome us, all yeah
With so many lightyears to go
And things to be found (to be found)
I'm sure that we'll all miss her so
[Hook]
It's the final countdown
The final countdown.
Guitar solo etc
It’s the final countdown